+++ manager/src/django/ppl_generator.py
