+++ tools/ppl_runner/README.md
@@ -1,0 +1,12 @@
+# Pipeline runner
+
+This folder contains a simple wrapper `ppl_runner.py` alongside with configuration files for testing `PipelineGenerator` and `PipelineConfigGenerator` Python classes that are used in production for dynamic pipeline configuration.
+
+## Usage
+
+The tool in the default configuration can be run by command `python ppl_runner.py`.
+
+
+
+Run `ppl_runner.py --help` for detailed information on configurability.
+
