+++ tools/ppl_runner/sample_camera_settings.json
